[Update from Kerson]

+++ prebuild/amba/include/amba_tvm.h
@@ -50,7 +50,7 @@ extern "C" {
 typedef struct {
 	IN const char* engine_name;
 	IN const char* engine_filepath;
-	INOUT uint32_t engine_id;
+	INOUT unsigned long engine_id;
 	uint32_t reserve[11];
 } amba_engine_cfg_t;
